Use fn ptr to print verbose messages.
Rather than check the verbose flag every time we want to print a message, set function pointers to point to either a dummy function (if -v or -V not set) or to a simple print function (if -v or -V are set).

Added -V flag to give 3 total verbosity levels.