Implement better template header functionality
Not happy with it, going to redo so that instead of being a free for all of tokens, each line represents 1 directive and all arguments for said directive are on the same line. Just makes more sense and will make it way easier to validate header.

Currently handling header in 2 passes:
1. Read file until closing header brace found
2. Process all the tokens

Shouldn't have to do this, should be able to process each lines tokens as it is read. Will find potential errors faster than reading the whole file.